backend/requirement.txt
absl-py @ file:///tmp/build/80754af9/absl-py_1615411205513/work
aiohttp @ file:///tmp/build/80754af9/aiohttp_1614360989480/work
aniso8601==9.0.1
astunparse==1.6.3
async-timeout==3.0.1
attrs @ file:///tmp/build/80754af9/attrs_1604765588209/work
beautifulsoup4 @ file:///home/linux1/recipes/ci/beautifulsoup4_1610988766420/work
blinker==1.4
brotlipy==0.7.0
cachetools @ file:///tmp/build/80754af9/cachetools_1611600262290/work
certifi==2020.12.5
cffi @ file:///tmp/build/80754af9/cffi_1613246939562/work
chardet @ file:///tmp/build/80754af9/chardet_1605303159953/work
click @ file:///home/linux1/recipes/ci/click_1610990599742/work
colorama==0.4.4
colorlog==4.8.0
coverage @ file:///tmp/build/80754af9/coverage_1614614864940/work
cryptography @ file:///tmp/build/80754af9/cryptography_1616769182610/work
cycler==0.10.0
Cython @ file:///tmp/build/80754af9/cython_1614014840177/work
decorator @ file:///tmp/build/80754af9/decorator_1621259047763/work
dnspython==2.1.0
et-xmlfile==1.0.1
filelock==3.0.12
Flask==1.1.1
Flask-RESTful==0.3.8
Flask-Script==2.0.6
flatbuffers @ file:///tmp/build/80754af9/python-flatbuffers_1614345733764/work
gast @ file:///tmp/build/80754af9/gast_1597433534803/work
gevent==1.4.0
google-auth @ file:///tmp/build/80754af9/google-auth_1616008050444/work
google-auth-oauthlib @ file:///tmp/build/80754af9/google-auth-oauthlib_1617120569401/work
google-pasta==0.2.0
greenlet @ file:///tmp/build/80754af9/greenlet_1620913644811/work
grpcio @ file:///tmp/build/80754af9/grpcio_1614883945333/work
h5py @ file:///tmp/build/80754af9/h5py_1593454119955/work
hashids==1.3.1
idna @ file:///home/linux1/recipes/ci/idna_1610986105248/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1615900449442/work
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
jieba==0.42.1
Jinja2 @ file:///tmp/build/80754af9/jinja2_1621238361758/work
joblib==1.0.1
Keras @ file:///tmp/build/80754af9/keras_split_1593112142734/work
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///tmp/build/80754af9/kiwisolver_1612282414123/work
logging==0.4.9.6
lxml @ file:///tmp/build/80754af9/lxml_1616442912239/work
Markdown @ file:///tmp/build/80754af9/markdown_1614363832606/work
MarkupSafe @ file:///tmp/build/80754af9/markupsafe_1621528142364/work
matplotlib==3.4.2
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multidict @ file:///tmp/build/80754af9/multidict_1607367738239/work
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603479632437/work
oauthlib==3.1.0
olefile==0.46
openpyxl==3.0.7
opt-einsum==3.3.0
pandas==1.2.3
Pillow @ file:///tmp/build/80754af9/pillow_1617386154241/work
protobuf==3.14.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///tmp/build/80754af9/pycparser_1594388511720/work
PyJWT==1.7.1
pymongo==3.11.3
PyMySQL @ file:///tmp/build/80754af9/pymysql_1610482900657/work
pyOpenSSL @ file:///tmp/build/80754af9/pyopenssl_1608057966937/work
pyparsing @ file:///home/linux1/recipes/ci/pyparsing_1610983426697/work
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil @ file:///home/ktietz/src/ci/python-dateutil_1611928101742/work
pythonping==1.0.16
pytz @ file:///tmp/build/80754af9/pytz_1612215392582/work
PyYAML==5.4.1
redis==3.5.3
requests @ file:///tmp/build/80754af9/requests_1608241421344/work
requests-file==1.5.1
requests-oauthlib==1.3.0
rsa @ file:///tmp/build/80754af9/rsa_1614366226499/work
scikit-learn==0.24.1
scipy @ file:///tmp/build/80754af9/scipy_1616700973326/work
seaborn @ file:///tmp/build/80754af9/seaborn_1608578541026/work
six @ file:///tmp/build/80754af9/six_1605205313296/work
sklearn==0.0
soupsieve @ file:///tmp/build/80754af9/soupsieve_1616183228191/work
tensorboard @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorboard_1614593728657/work/tmp_pip_dir
tensorboard-plugin-wit==1.6.0
tensorflow @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_cpu/tensorflow-base_1614346860851/work/tensorflow_pkg/tensorflow-2.4.1-cp37-cp37m-linux_x86_64.whl
tensorflow-estimator @ file:///home/builder/ktietz/aggregate/tensorflow_recipes/ci_te/tensorflow-estimator_1614592767119/work/whl_temp/tensorflow_estimator-2.4.0-py2.py3-none-any.whl
termcolor==1.1.0
threadpoolctl==2.1.0
tldextract==3.1.0
tornado @ file:///tmp/build/80754af9/tornado_1606942283357/work
typing-extensions @ file:///tmp/build/80754af9/typing_extensions_1611751222202/work
urllib3 @ file:///tmp/build/80754af9/urllib3_1615837158687/work
validators @ file:///tmp/build/80754af9/validators_1612286467315/work
Werkzeug @ file:///home/ktietz/src/ci/werkzeug_1611932622770/work
wrapt==1.12.1
yarl @ file:///tmp/build/80754af9/yarl_1606939952652/work
zipp @ file:///tmp/build/80754af9/zipp_1615904174917/work
zoomeye==2.0.4.6.5
